Fixed block size (ranges from 0 to 1 instead of -1 to 1). Also fixed a stupid depth buffer error (it was not bound :) )

+++ VoxelEngine/VoxelEngine/Core/Chunk.h
@@ -16,7 +16,7 @@ public:
 
 	const uint32_t GetID();
 
-	const Block* GetBlock(uint8_t x, uint8_t y, uint8_t z);
+	const Block* GetBlock(unsigned int x, unsigned int y, unsigned int z);
 
 	void SetActive(const bool active);
 	const bool GetActive();
